Change directories for pages and change types for new pages
