+++ src/main.jsx
@@ -1,11 +1,11 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import "./index.css";
-import App from "./App.jsx";
-import Generator from "./Generator.jsx";
+import App from "./pages/Main/App.jsx";
+import Generator from "./pages/Generator/Generator.tsx";
 import { HashRouter, Routes, Route } from "react-router";
-import Login from "./Login.jsx";
-import Settings from "./Settings.jsx";
+import Login from "./pages/Login/Login.tsx";
+import Settings from "./pages/Settings/Settings.tsx";
 createRoot(document.getElementById("root")).render(
   <StrictMode>
     <HashRouter type="slash">
