+++ src/pages/Main/App.jsx
@@ -1,12 +1,9 @@
 // import { useState } from 'react'
-import reactLogo from "./assets/react.svg";
+import reactLogo from "../../assets/react.svg";
 import viteLogo from "/vite.svg";
 import "./App.css";
 import { NavLink } from "react-router";
-
 function App() {
-  // const [count, setCount] = useState(0)
-
   return (
     <>
       <div>
@@ -31,16 +28,7 @@ function App() {
         <NavLink className="nav-link" to="/user-settings">
           <button> Press Here to link to settings page</button>
         </NavLink>
-        {/* <button onClick={() => setCount((count) => count + 1)}>
-          count is {count}
-        </button> */}
-        {/* <p>
-          Edit <code>src/App.jsx</code> and save to test HMR
-        </p> */}
       </div>
-      {/* <p className="read-the-docs">
-        Click on the Vite and React logos to learn more
-      </p> */}
     </>
   );
 }
